refactor(ui): added void tool at index 0.
when the user starts, they won't modify the model inadvertently.
void is a 'safe' tool that allows the user to explore the model without inadvertently modifying it.

escape last level resets the tools index positioning the user to the void tool

space last level keeps the user on the current tool